@@ -15,7 +15,7 @@ void main() {
       'plugins.flutter.io/shared_preferences',
     );
 
-    const Map<String, dynamic> kTestValues = <String, dynamic>{
+    const Map<String, Object> kTestValues = <String, Object>{
       'flutter.String': 'hello world',
       'flutter.Bool': true,
       'flutter.Int': 42,
@@ -23,10 +23,10 @@ void main() {
       'flutter.StringList': <String>['foo', 'bar'],
     };
 
-    InMemorySharedPreferencesStore testData;
+    late InMemorySharedPreferencesStore testData;
 
     final List<MethodCall> log = <MethodCall>[];
-    MethodChannelSharedPreferencesStore store;
+    late MethodChannelSharedPreferencesStore store;
 
     setUp(() async {
       testData = InMemorySharedPreferencesStore.empty();
@@ -44,9 +44,9 @@ void main() {
           return await testData.clear();
         }
         final RegExp setterRegExp = RegExp(r'set(.*)');
-        final Match match = setterRegExp.matchAsPrefix(methodCall.method);
-        if (match.groupCount == 1) {
-          final String valueType = match.group(1);
+        final Match? match = setterRegExp.matchAsPrefix(methodCall.method);
+        if (match?.groupCount == 1) {
+          final String valueType = match!.group(1)!;
           final String key = methodCall.arguments['key'];
           final Object value = methodCall.arguments['value'];
           return await testData.setValue(valueType, key, value);
@@ -59,8 +59,6 @@ void main() {
 
     tearDown(() async {
       await testData.clear();
-      store = null;
-      testData = null;
     });
 
     test('getAll', () async {
